add ability to pass prior run data in context
Adds a new `result.Result` struct that wraps Go's error interface and
allows Spec subclasses to pass back arbitrary data representing the
results of the Run that are then stored in the context that is supplied
to the next Run() call and can be used by plugins to query prior run
results.

This will be used by the gdt-http and gdt-kube plugins.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>