make timeout handling more robust
Adds a new gdttypes.Timeout struct with fields for the timeout duration
and whether or not to expect the timeout to occur. The latter expect
field is useful in unit testing the timeout functionality.

Adds a `gdtcontext.TimedOut()` function which examines a context and an
arbitrary error to determine if the context is timed out. This manifests
as either a context deadline exceeded error or, depending on the OS, a
signal: killed error when os/exec.CmdContext() is used.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>