refactor assertion handling in exec specs
Adds a new ErrFailure base error class to the `gdterrors` package along
with a few helper functions (`NotEqual`, `NotIn` and `NoneIn`) that
return subclasses of ErrFailure when an assertion fails.

Refactors the exec plugin's Spec.Run() to not call testify's assert or
require package but instead call the new `assertions.OK()` method to
determine if all assertions passed, and if not, call
`testing.T:Error()`. This allows us to retry the executed command if the
assertions did not fail with a terminal condition.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>