add standard Spec.Wait configuration
Adds a new `wait` field to the base `gdttypes.Spec` definition. This
`wait` field is a `gdttypes.Wait` struct that contains a `before` and
`after` field which are optional strings that contains time string
durations for gdt to wait before and after executing the test spec
action.

In action, this is what this new `wait` configuration looks like:

```yaml
name: echo-wait
description: a scenario that echoes the word "cat" and has some waits configured
tests:
  - name: wait-before
    exec: echo "cat"
    wait:
      before: 10ms
  - name: wait-after
    exec: echo "cat"
    wait:
      after: 20ms
```

Signed-off-by: Jay Pipes <jaypipes@gmail.com>