bring in standardized Assertions type
Brings in a standardized `gdttypes.Assertions` interface that provides
some methods for asserting that expected conditions were met about a
result or response from an action a test spec takes.

Updates the `exec` plugin to follow the new `gdttypes.Assertions`
interface in preparation for creation of a new `assertion` package that
will contain standardized versions of, for example, the `JSONAssertions`
object from the `gdt-http` plugin which will be used in the `gdt-kube`
plugin.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>